+++ src/Server/SovereignServer/Data/Scripts/Sovereign/Motd.lua
@@ -1,6 +1,11 @@
+g_ticks = 0
+
+function on_tick()
+    if g_ticks % 50 == 0 then
+        util.log_info("tick " .. g_ticks)
+    end
+    g_ticks = g_ticks + 1
+end
+
 util.log_info("This is a test of the scripting engine.")
-util.log_warn("This is a warning.")
-util.log_error("This is an error.")
-util.log_crit("This is a critical message.")
-util.log_trace("This is a trace message.")
-util.log_debug("This is a debug message.")
+scripting.add_event_callback(events.Core_Tick, on_tick)